refactor(tests): Move common test setup into integration test dir.
